rename useGlobal and withGlobal to their plural counterparts
